+++ CHANGELOG.md
@@ -12,6 +12,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add more tests.
 - Add `JavaDoc`.
 
+## [1.3.0](https://github.com/appulse-projects/utils-java/releases/tag/1.3.0) - 2018-02-01
+
+Add resource utils class and made small refactorign.
+
+### Added
+
+- `ResourceUtils` which contains resource related functions.
+
+### Changed
+
+- Renamed `BytesUtil` to `BytesUtils`.
+- Moved methods for reading `InputStream` to `BytesUtils`.
+
 ## [1.2.1](https://github.com/appulse-projects/utils-java/releases/tag/1.2.1) - 2018-02-01
 
 Add socket utils class, with help functions related to sockets work.